+++ cmd/buf/internal/command/breaking/breaking.go
@@ -53,6 +53,7 @@ const (
 	againstRegistryFlagName   = "against-registry"
 	excludePathsFlagName      = "exclude-path"
 	disableSymlinksFlagName   = "disable-symlinks"
+	lockedFlagName            = "locked"
 )
 
 // NewCommand returns a new Command.
@@ -92,6 +93,7 @@ type flags struct {
 	AgainstRegistry   bool
 	ExcludePaths      []string
 	DisableSymlinks   bool
+	Locked            bool
 	// special
 	InputHashtag string
 }
@@ -105,6 +107,7 @@ func (f *flags) Bind(flagSet *pflag.FlagSet) {
 	bufcli.BindInputHashtag(flagSet, &f.InputHashtag)
 	bufcli.BindExcludePaths(flagSet, &f.ExcludePaths, excludePathsFlagName)
 	bufcli.BindDisableSymlinks(flagSet, &f.DisableSymlinks, disableSymlinksFlagName)
+	bufcli.BindLocked(flagSet, &f.Locked, lockedFlagName)
 	flagSet.StringVar(
 		&f.ErrorFormat,
 		errorFormatFlagName,
@@ -201,6 +204,7 @@ func run(
 		wasmRuntime,
 		bufctl.WithTargetPaths(flags.Paths, flags.ExcludePaths),
 		bufctl.WithConfigOverride(flags.Config),
+		bufctl.WithLocked(flags.Locked),
 	)
 	if err != nil {
 		return err